refactor: Move difficulty page content to common/DeckListPage
