+++ frontend/src/components/common/Difficulty.module.scss
